Use fixed-length names for prepared queries (uuids), fix problem with close.
Calling (connection close) was causing the underlying postgres connection
to be released twice: once on the close and once when the connection
object was released.

I had previously used the query text as names of prepared statements but
had some problems that I suspected were due to name collisions for
overly-long queries. To avoid that, I started using uuids as query names
and keeping the query-to-uuid map as a dictionary in the connection
object.